fix: Hyperlink Styling for changed Themes
