+++ src/TypeDefs.h
@@ -318,6 +318,8 @@ typedef struct _globals_t
 
   WCHAR     WorkingDirectory[MAX_PATH];
   WCHAR     IniFile[MAX_PATH];
+  WCHAR     IniFileDefault[MAX_PATH];
+
   WCHAR     CurrentFile[MAX_PATH];
 
 } GLOBALS_T, *PGLOBALS_T;
