+++ themes/style_schemata/Dark Theme.ini
